[team1] 다미 수정사항 author: 김다미 file: teamBootstrap/src/main/resources/templates/production/addProductionPlan.html teamBootstrap/src/main/java/ksmart39/springboot/controller/ProductionController.java teamBootstrap/src/main/java/ksmart39/springboot/dao/ProductionMapper.java teamBootstrap/src/main/java/ksmart39/springboot/dao/ProductionPlanMapper.java teamBootstrap/src/main/resources/mapper/ProductionPlanMapper.xml teamBootstrap/src/main/java/ksmart39/springboot/service/ProductionPlanService.java teamBootstrap/src/main/java/ksmart39/springboot/service/ProductionService.java detail: 생산계획 mapper, service, xml 파일 생성
