+++ teamBootstrap/src/main/java/ksmart39/springboot/service/ProductionService.java
@@ -15,17 +15,5 @@ public class ProductionService {
 	@Autowired
 	private ProductionMapper productionMapper;
 	
-	//[다미]소분류 조회
-	public List<Map<String, Object>> getDetailCode(String gender){
-		return productionMapper.getDetailCode(gender);
-	}
-	
-	//[다미]생산코드 조회
-	public List<Map<String, Object>> getProductCode(){
-		return productionMapper.getProductCode();
-	}	
-	
-	public List<ProductionPlan> getProductionMonthlyPlanList(){
-		return productionMapper.getProductionMonthlyPlanList();
-	}
+
 }
