+++ teamBootstrap/src/main/resources/mapper/ProductionPlanMapper.xml
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="ksmart39.springboot.dao.ProductionPlanMapper">
+
+
+<select id="getProductionMonthlyPlanList" resultType="ProductionPlan">
+		SELECT 
+			production_plan_code
+			,gender_categorized_name
+			,detailed_categorized_name
+			,charge_employee_code
+			,charge_employee_name
+			,production_plan_quantity
+			,production_plan_qantity_unit
+			,expected_production_start_date
+			,expected_production_end_date
+			,production_plan_period
+			,production_plan_note
+			,production_plan_reg_date
+			,production_plan_update_date
+		FROM 
+			production_plan;
+</select>
+
+	<select id="getDetailCode" parameterType="String" resultType="Map">
+		SELECT 
+		   detailed_categorized_name
+		   ,product_code
+		FROM 
+			product_code_detail
+      	WHERE 
+			gender_categorized_name = #{gender};
+	</select>
+	
+	<select id="getProductCode" resultType="Map">
+		SELECT 
+			DISTINCT(gender_categorized_name)
+		FROM 
+			product_code_detail;
+	</select>
+</mapper>
